refactor: split sarin.go into runner.go and worker.go
